encapsuling draft component
