+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,3 +1,5 @@
+# Pull Request
+
 ## Description
 
 Please describe the changes made in this pull request.
