+++ src/pages/draft/index.tsx
@@ -1,0 +1,5 @@
+import DraftContainer from '@/components/DraftContainer';
+
+export default function Draft() {
+  return <DraftContainer />
+}
