+++ src/styles/globals.css
@@ -21,5 +21,5 @@ body {
   background: '#fff';
   background-image: url('/background.jpg');
   background-size: cover;
-  background-position: center;
+  background-repeat: no-repeat;
 }
